+++ crates/sdk-core/src/worker/mod.rs
@@ -319,16 +319,16 @@ impl WorkerConfig {
 
     pub(crate) fn computed_deployment_version(&self) -> Option<WorkerDeploymentVersion> {
         let wdv = match self.versioning_strategy {
-            WorkerVersioningStrategy::None { ref build_id } => WorkerDeploymentVersion {
-                deployment_name: "".to_owned(),
-                build_id: build_id.clone(),
-            },
+            WorkerVersioningStrategy::None { ref build_id } => WorkerDeploymentVersion::builder()
+                .deployment_name("")
+                .build_id(build_id.clone())
+                .build(),
             WorkerVersioningStrategy::WorkerDeploymentBased(ref opts) => opts.version.clone(),
             WorkerVersioningStrategy::LegacyBuildIdBased { ref build_id } => {
-                WorkerDeploymentVersion {
-                    deployment_name: "".to_owned(),
-                    build_id: build_id.clone(),
-                }
+                WorkerDeploymentVersion::builder()
+                    .deployment_name("")
+                    .build_id(build_id.clone())
+                    .build()
             }
         };
         if wdv.is_empty() { None } else { Some(wdv) }
@@ -2779,10 +2779,12 @@ mod tests {
             .namespace("default")
             .task_queue("test-queue")
             .versioning_strategy(WorkerVersioningStrategy::WorkerDeploymentBased(
-                WorkerDeploymentOptions::new(WorkerDeploymentVersion {
-                    deployment_name: "deployment".to_string(),
-                    build_id: "1.0".to_string(),
-                })
+                WorkerDeploymentOptions::new(
+                    WorkerDeploymentVersion::builder()
+                        .deployment_name("deployment")
+                        .build_id("1.0")
+                        .build(),
+                )
                 .default_versioning_behavior(VersioningBehavior::AutoUpgrade.into())
                 .build(),
             ))