+++ crates/sdk-core/tests/integ_tests/workflow_tests.rs
@@ -558,16 +558,20 @@ async fn deployment_version_correct_in_wf_info(#[values(true, false)] use_only_b
     let wf_type = "deployment_version_correct_in_wf_info";
     let mut starter = CoreWfStarter::new(wf_type);
     starter.sdk_config.deployment_options = if use_only_build_id {
-        WorkerDeploymentOptions::new(WorkerDeploymentVersion {
-            deployment_name: "".to_string(),
-            build_id: "1.0".to_string(),
-        })
+        WorkerDeploymentOptions::new(
+            WorkerDeploymentVersion::builder()
+                .deployment_name("".to_string())
+                .build_id("1.0".to_string())
+                .build(),
+        )
         .build()
     } else {
-        WorkerDeploymentOptions::new(WorkerDeploymentVersion {
-            deployment_name: "deployment-1".to_string(),
-            build_id: "1.0".to_string(),
-        })
+        WorkerDeploymentOptions::new(
+            WorkerDeploymentVersion::builder()
+                .deployment_name("deployment-1".to_string())
+                .build_id("1.0".to_string())
+                .build(),
+        )
         .build()
     };
     starter.set_core_task_types(WorkerTaskTypes::workflow_only());
@@ -673,16 +677,20 @@ async fn deployment_version_correct_in_wf_info(#[values(true, false)] use_only_b
 
     let mut starter = starter.clone_no_worker();
     starter.sdk_config.deployment_options = if use_only_build_id {
-        WorkerDeploymentOptions::new(WorkerDeploymentVersion {
-            deployment_name: "".to_string(),
-            build_id: "2.0".to_string(),
-        })
+        WorkerDeploymentOptions::new(
+            WorkerDeploymentVersion::builder()
+                .deployment_name("".to_string())
+                .build_id("2.0".to_string())
+                .build(),
+        )
         .build()
     } else {
-        WorkerDeploymentOptions::new(WorkerDeploymentVersion {
-            deployment_name: "deployment-1".to_string(),
-            build_id: "2.0".to_string(),
-        })
+        WorkerDeploymentOptions::new(
+            WorkerDeploymentVersion::builder()
+                .deployment_name("deployment-1".to_string())
+                .build_id("2.0".to_string())
+                .build(),
+        )
         .build()
     };